requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.9
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
kombu==5.2.4
packaging==21.3
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.30
psutil==5.9.4
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-decouple==3.6
python-ffmpeg-video-streaming==0.1.15
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tornado==6.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
